Reduce use-case orchestrator fan-out
Refactor diagnostics and recording orchestration into focused helper modules.

- extract summary/result and peak-finding helpers
- slim RunRecorder with recorder runtime/types helpers
- restore completed history insights status contract for live polling
- update tests and pipeline docs for the moved private seams
