+++ apps/server/vibesensor/use_cases/diagnostics/summary_builder.py
@@ -3,40 +3,15 @@
 from __future__ import annotations
 
 from collections.abc import Callable
-from dataclasses import dataclass, replace
 from pathlib import Path
-from statistics import median as _median
-from typing import cast
+from typing import TYPE_CHECKING
 
-from vibesensor.domain import (
-    ConfigurationSnapshot,
-    DiagnosticCase,
-    DrivingPhaseInterval,
-    LocationIntensitySummary,
-    RunCapture,
-    RunSetup,
-    RunSuitability,
-    Sensor,
-    SpeedSource,
-    TestRun,
-)
 from vibesensor.domain import (
     Finding as DomainFinding,
 )
-from vibesensor.domain.test_plan import plan_test_actions
 from vibesensor.domain.vibration_origin import VibrationOrigin
 from vibesensor.report_i18n import normalize_lang
 from vibesensor.shared.boundaries.analysis_payload import AnalysisSummary
-from vibesensor.shared.boundaries.diagnostic_case import (
-    case_context_from_metadata,
-)
-from vibesensor.shared.boundaries.finding import step_payloads_from_plan
-from vibesensor.shared.boundaries.summary_serialization import (
-    build_summary_payload,
-    serialize_plot_data,
-)
-from vibesensor.shared.run_context import build_summary_warnings
-from vibesensor.shared.time_utils import utc_now_iso
 from vibesensor.shared.types.json_types import JsonObject
 from vibesensor.use_cases.diagnostics._types import (
     AccelStatistics,
@@ -47,169 +22,28 @@ from vibesensor.use_cases.diagnostics.helpers import (
     _load_run,
     _validate_required_strength_metrics,
 )
-from vibesensor.use_cases.diagnostics.phase_segmentation import (
-    DrivingPhase,
-)
-from vibesensor.use_cases.diagnostics.plots import _plot_data
 from vibesensor.use_cases.diagnostics.run_data_preparation import (
     PreparedRunData,
-    build_domain_driving_segments,
-    build_phase_summary,
-    build_phase_timeline,
-    build_sensor_analysis,
     prepare_run_data,
 )
-from vibesensor.use_cases.diagnostics.speed_profile_helpers import (
-    _speed_stats,
-)
-from vibesensor.use_cases.diagnostics.statistics import (
-    _strength_band_key,
-    compute_accel_statistics,
-    compute_frame_integrity_counts,
-    compute_reference_completeness,
-)
-from vibesensor.use_cases.diagnostics.top_cause_selection import select_top_causes
+from vibesensor.use_cases.diagnostics.statistics import compute_accel_statistics
+
+from . import _summary_result, _summary_steps
+
+if TYPE_CHECKING:
+    from vibesensor.domain import TestRun
 
 # ═══════════════════════════════════════════════════════════════════════════
 # Origin helpers
 # ═══════════════════════════════════════════════════════════════════════════
 
 
-def summarize_origin(
-    findings: tuple[DomainFinding, ...],
-) -> VibrationOrigin | None:
+def summarize_origin(findings: tuple[DomainFinding, ...]) -> VibrationOrigin | None:
     """Return the most-likely origin as a domain value object."""
     return VibrationOrigin.from_ranked_findings(findings)
 
 
-def build_findings_bundle(
-    metadata: JsonObject,
-    samples: list[Sample],
-    *,
-    language: str,
-    prepared: PreparedRunData,
-    overall_strength_band_key: str | None,
-    has_reference_gaps: bool,
-    sensor_count: int,
-    findings_builder: Callable[..., tuple[DomainFinding, ...]] | None = None,
-) -> tuple[
-    VibrationOrigin | None,
-    list[DrivingPhaseInterval],
-    tuple[DomainFinding, ...],
-    tuple[DomainFinding, ...],
-]:
-    """Build findings plus derived diagnosis narrative fields.
-
-    Returns ``(origin, timeline, domain_findings, domain_top_causes)``.
-    Findings are returned with :class:`ConfidenceAssessment` already
-    attached via :meth:`Finding.with_confidence_assessment`.
-    """
-    builder = findings_builder or _build_findings
-    domain_findings = builder(
-        metadata=metadata,
-        samples=samples,
-        speed_sufficient=prepared.speed_sufficient,
-        steady_speed=prepared.is_steady_speed,
-        speed_stddev_kmh=prepared.speed_stddev_kmh,
-        speed_non_null_pct=prepared.speed_non_null_pct,
-        raw_sample_rate_hz=prepared.raw_sample_rate_hz,
-        lang=language,
-        per_sample_phases=prepared.per_sample_phases,
-        run_noise_baseline_g=prepared.run_noise_baseline_g,
-    )
-
-    # Enrich findings with ConfidenceAssessment at construction time
-    domain_findings = tuple(
-        f
-        if f.confidence_assessment is not None
-        else f.with_confidence_assessment(
-            strength_band_key=overall_strength_band_key or "",
-            steady_speed=prepared.is_steady_speed,
-            has_reference_gaps=has_reference_gaps,
-            sensor_count=sensor_count,
-        )
-        for f in domain_findings
-    )
-
-    domain_diagnostic_findings = tuple(f for f in domain_findings if not f.is_reference)
-    most_likely_origin = summarize_origin(
-        domain_diagnostic_findings,
-    )
-    phase_timeline = build_phase_timeline(
-        prepared.phase_segments,
-        domain_findings,
-        min_confidence=0.25,
-    )
-    domain_top_causes = select_top_causes(
-        domain_findings,
-    )
-    return (
-        most_likely_origin,
-        phase_timeline,
-        domain_findings,
-        domain_top_causes,
-    )
-
-
-def build_sensor_bundle(
-    samples: list[Sample],
-    *,
-    language: str,
-    per_sample_phases: list[DrivingPhase],
-) -> tuple[list[str], set[str], list[LocationIntensitySummary]]:
-    """Build location-scoped sensor summaries used by analysis and reports."""
-    return build_sensor_analysis(
-        samples=samples,
-        language=language,
-        per_sample_phases=per_sample_phases,
-    )
-
-
-def build_run_suitability_bundle(
-    metadata: JsonObject,
-    samples: list[Sample],
-    *,
-    prepared: PreparedRunData,
-    accel_stats: AccelStatistics,
-) -> tuple[bool, RunSuitability | None, str | None]:
-    """Build run-suitability checks and related confidence context."""
-    reference_complete = compute_reference_completeness(metadata)
-    sensor_ids = {
-        str(cid)
-        for sample in samples
-        if isinstance(sample, dict) and (cid := sample.get("client_id"))
-    }
-    total_dropped, total_overflow = compute_frame_integrity_counts(samples)
-    run_suitability = RunSuitability.evaluate(
-        steady_speed=prepared.is_steady_speed,
-        speed_sufficient=prepared.speed_sufficient,
-        sensor_count=len(sensor_ids),
-        reference_complete=reference_complete,
-        sat_count=accel_stats["sat_count"],
-        total_dropped=total_dropped,
-        total_overflow=total_overflow,
-    )
-    amp_metric_values = accel_stats["amp_metric_values"]
-    overall_strength_band_key = (
-        _strength_band_key(_median(amp_metric_values)) if amp_metric_values else None
-    )
-    return reference_complete, run_suitability, overall_strength_band_key
-
-
-@dataclass(frozen=True, slots=True)
-class AnalysisResult:
-    """Output coordinator: carries domain aggregates alongside the boundary summary dict.
-
-    Returned by :meth:`RunAnalysis.summarize`.  The ``summary`` dict is
-    still needed for persistence (SQLite stores it as JSON) and many
-    existing boundary consumers.  ``test_run`` and ``diagnostic_case``
-    expose the fully-constructed domain aggregates so that callers no
-    longer need to discard them.
-    """
-
-    test_run: TestRun
-    diagnostic_case: DiagnosticCase
-    summary: AnalysisSummary
+AnalysisResult = _summary_result.AnalysisResult
 
 
 class RunAnalysis:
@@ -288,24 +122,26 @@ class RunAnalysis:
         ``summary`` dict.
         """
         reference_complete, run_suitability, overall_strength_band_key = (
-            build_run_suitability_bundle(
+            _summary_steps.build_run_suitability_bundle(
                 self._metadata,
                 self._samples,
                 prepared=self._prepared,
                 accel_stats=self._accel_stats,
             )
         )
-        sensor_locations, connected_locations, sensor_intensity_by_location = build_sensor_bundle(
-            self._samples,
-            language=self._language,
-            per_sample_phases=self._prepared.per_sample_phases,
+        sensor_locations, connected_locations, sensor_intensity_by_location = (
+            _summary_steps.build_sensor_bundle(
+                self._samples,
+                language=self._language,
+                per_sample_phases=self._prepared.per_sample_phases,
+            )
         )
         (
             most_likely_origin,
             phase_timeline,
             domain_findings,
             domain_top_causes,
-        ) = build_findings_bundle(
+        ) = _summary_steps.build_findings_bundle(
             self._metadata,
             self._samples,
             language=self._language,
@@ -315,120 +151,26 @@ class RunAnalysis:
             sensor_count=len(sensor_locations),
             findings_builder=self._findings_builder,
         )
-
-        # Build the domain aggregate with run-level value objects
-        speed_profile = self._prepared.speed_profile if self._prepared.speed_values else None
-        domain_suitability = run_suitability
-
-        # Derive top_causes as a subset of the enriched findings,
-        # preserving signatures collected by group_findings_by_source
-        top_cause_ids = {f.finding_id for f in domain_top_causes if f.finding_id}
-        top_cause_sigs = {f.finding_id: f.signatures for f in domain_top_causes if f.signatures}
-        final_top_causes_list: list[DomainFinding] = []
-        for f in domain_findings:
-            if f.finding_id in top_cause_ids:
-                sigs = top_cause_sigs.get(f.finding_id)
-                final_top_causes_list.append(replace(f, signatures=sigs) if sigs else f)
-        final_top_causes = tuple(final_top_causes_list)
-
-        configuration_snapshot = ConfigurationSnapshot.from_metadata(self._metadata)
-        driving_segments = build_domain_driving_segments(self._prepared.phase_segments)
-        domain_test_plan = plan_test_actions(domain_findings)
-        summary_test_plan: list[JsonObject] = cast(
-            list[JsonObject], step_payloads_from_plan(domain_test_plan)
-        )
-        _raw_settings = self._metadata.get("analysis_settings")
-        _scalar_settings: list[tuple[str, int | float | bool | str]] = []
-        if isinstance(_raw_settings, dict):
-            for _k, _v in sorted(_raw_settings.items()):
-                if isinstance(_v, (int, float, bool, str)):
-                    _scalar_settings.append((_k, _v))
-        capture = RunCapture(
-            run_id=self._prepared.run_id,
-            setup=RunSetup(
-                sensors=Sensor.from_location_codes(sensor_locations) if sensor_locations else (),
-                speed_source=SpeedSource(),
-                configuration_snapshot=configuration_snapshot,
-            ),
-            analysis_settings=tuple(_scalar_settings),
-            sample_count=len(self._samples),
-            duration_s=self._prepared.duration_s,
-        )
-        self._test_run = TestRun(
-            capture=capture,
-            driving_segments=driving_segments,
-            findings=domain_findings,
-            top_causes=final_top_causes,
-            speed_profile=speed_profile,
-            suitability=domain_suitability,
-            test_plan=domain_test_plan,
-        )
-        domain_car, domain_symptoms = case_context_from_metadata(self._metadata)
-        diagnostic_case = DiagnosticCase.start(
-            car=domain_car,
-            symptoms=domain_symptoms,
-            test_plan=domain_test_plan,
-        ).add_run(self._test_run)
-
-        summary_speed_stats = _speed_stats(self._prepared.speed_values)
-        summary_phase_info = build_phase_summary(self._prepared.phase_segments)
-
-        summary = build_summary_payload(
+        result = _summary_result.build_analysis_result(
             file_name=self._file_name,
-            run_id=self._prepared.run_id,
-            samples=self._samples,
-            duration_s=self._prepared.duration_s,
-            language=self._language,
             metadata=self._metadata,
-            raw_sample_rate_hz=self._prepared.raw_sample_rate_hz,
-            speed_breakdown=self._prepared.speed_breakdown,
-            phase_speed_breakdown=self._prepared.phase_speed_breakdown,
-            phase_segments=self._prepared.phase_segments,
-            run_noise_baseline_g=self._prepared.run_noise_baseline_g,
-            speed_breakdown_skipped_reason=self._prepared.speed_breakdown_skipped_reason,
-            findings=domain_findings,
-            top_causes=final_top_causes,
-            most_likely_origin=most_likely_origin,
-            test_plan=summary_test_plan,
-            phase_timeline=phase_timeline,
-            speed_stats=summary_speed_stats,
-            speed_stats_by_phase=self._prepared.speed_stats_by_phase,
-            phase_info=summary_phase_info,
+            samples=self._samples,
+            language=self._language,
+            include_samples=self._include_samples,
+            prepared=self._prepared,
+            accel_stats=self._accel_stats,
             sensor_locations=sensor_locations,
             connected_locations=connected_locations,
             sensor_intensity_by_location=sensor_intensity_by_location,
-            run_suitability=domain_suitability,
-            speed_values=self._prepared.speed_values,
-            speed_non_null_pct=self._prepared.speed_non_null_pct,
-            accel_stats=self._accel_stats,
-            amp_metric_values=self._accel_stats["amp_metric_values"],
-        )
-        summary["warnings"] = build_summary_warnings(
-            self._metadata,
             reference_complete=reference_complete,
+            run_suitability=run_suitability,
+            most_likely_origin=most_likely_origin,
+            phase_timeline=phase_timeline,
+            domain_findings=domain_findings,
+            domain_top_causes=domain_top_causes,
         )
-        summary["report_date"] = self._metadata.get("end_time_utc") or utc_now_iso()
-        summary["plots"] = serialize_plot_data(
-            _plot_data(
-                samples=self._samples,
-                speed_breakdown=self._prepared.speed_breakdown,
-                phase_speed_breakdown=self._prepared.phase_speed_breakdown,
-                findings=domain_findings,
-                raw_sample_rate_hz=self._prepared.raw_sample_rate_hz,
-                steady_speed=self._prepared.is_steady_speed,
-                run_noise_baseline_g=self._prepared.run_noise_baseline_g,
-                per_sample_phases=self._prepared.per_sample_phases,
-                phase_segments=self._prepared.phase_segments,
-            ),
-        )
-        cast(dict[str, object], summary)["_summary_version"] = 2
-        if not self._include_samples:
-            summary.pop("samples", None)
-        return AnalysisResult(
-            test_run=self._test_run,
-            diagnostic_case=diagnostic_case,
-            summary=summary,
-        )
+        self._test_run = result.test_run
+        return result
 
 
 def summarize_run_data(
@@ -480,6 +222,7 @@ def build_findings_for_samples(
         raw_sample_rate_hz=prepared.raw_sample_rate_hz,
         lang=language,
         per_sample_phases=prepared.per_sample_phases,
+        run_noise_baseline_g=prepared.run_noise_baseline_g,
     )
