+++ apps/server/vibesensor/use_cases/run/logger.py
@@ -1,34 +1,14 @@
-"""Metrics recording orchestrator.
-
-``RunRecorder`` coordinates the recording loop, lifecycle delegation, and
-history-DB persistence. Focused helpers live in:
-
-- :mod:`vibesensor.use_cases.run.persistence_writer` — history-write
-  coordination, retry/backoff handling, and persistence health state.
-- :mod:`vibesensor.use_cases.run.sample_flush` — sample building, flush
-  decisions, and auto-stop timing checks.
-- :mod:`vibesensor.use_cases.run.lifecycle_state` — in-memory recording
-  session lifecycle state.
-- :mod:`vibesensor.use_cases.run.sample_builder` — pure sample record
-  construction.
-- :mod:`vibesensor.use_cases.run.post_analysis` — background analysis
-  thread/queue plus the injected post-stop analysis boundary.
-- :mod:`vibesensor.use_cases.run.status_reporting` — status and health
-  payload assembly above persistence/post-analysis collaborators.
-"""
+"""Thin recording orchestrator around the focused run helpers."""
 
 from __future__ import annotations
 
-import asyncio
 import logging
 import time
 from collections.abc import Callable
-from dataclasses import dataclass
 from threading import RLock
+from typing import TYPE_CHECKING
 from uuid import uuid4
 
-from vibesensor.domain.snapshots import AnalysisSettingsSnapshot
-from vibesensor.shared.constants import NUMERIC_TYPES
 from vibesensor.shared.ports import (
     ClientTracker,
     RunPersistence,
@@ -37,9 +17,7 @@ from vibesensor.shared.ports import (
     SpeedProvider,
 )
 from vibesensor.shared.time_utils import utc_now_iso
-from vibesensor.shared.types.health_snapshot import RunRecorderHealthSnapshot
-from vibesensor.shared.types.json_types import JsonObject
-from vibesensor.use_cases.run.lifecycle_state import ActiveRunSnapshot, RunLifecycleState
+from vibesensor.use_cases.run.lifecycle_state import RunLifecycleState
 from vibesensor.use_cases.run.persistence_writer import (
     _APPEND_RETRY_DELAYS_S,
     _MAX_APPEND_RETRIES,
@@ -48,162 +26,28 @@ from vibesensor.use_cases.run.persistence_writer import (
     RunPersistenceWriter,
 )
 from vibesensor.use_cases.run.post_analysis import PostAnalysisWorker, build_post_analysis_summary
-from vibesensor.use_cases.run.sample_builder import build_run_metadata, firmware_version_for_run
 from vibesensor.use_cases.run.sample_flush import SampleFlushOrchestrator
 from vibesensor.use_cases.run.status_reporting import (
     build_run_recorder_health_snapshot,
     build_run_recorder_status,
 )
 
+from . import _recorder_runtime, _recorder_types
+
+if TYPE_CHECKING:
+    from vibesensor.domain.snapshots import AnalysisSettingsSnapshot
+    from vibesensor.shared.types.health_snapshot import RunRecorderHealthSnapshot
+    from vibesensor.use_cases.run.lifecycle_state import ActiveRunSnapshot
+
 LOGGER = logging.getLogger(__name__)
 
 __all__ = [
-    "RecorderShutdownReport",
     "RunRecorder",
-    "RunRecorderConfig",
     "_APPEND_RETRY_DELAYS_S",
     "_MAX_APPEND_RETRIES",
     "_MAX_HISTORY_CREATE_RETRIES",
     "_RETRY_COOLDOWN_BASE_S",
 ]
-
-_DB_THREAD_TIMEOUT_S: float = 10.0
-"""Timeout for DB-bound asyncio.to_thread() calls; prevents a stalled DB from
-blocking the metrics-logger event loop indefinitely."""
-
-
-# ---------------------------------------------------------------------------
-# Configuration and shutdown report
-# ---------------------------------------------------------------------------
-
-
-@dataclass
-class RunRecorderConfig:
-    """Static configuration bundle for :class:`RunRecorder`."""
-
-    metrics_log_hz: int
-    sensor_model: str
-    default_sample_rate_hz: int
-    fft_window_size_samples: int
-    accel_scale_g_per_lsb: float | None = None
-    persist_history_db: bool = True
-    no_data_timeout_s: float = 15.0
-
-
-@dataclass(frozen=True, slots=True)
-class RecorderShutdownReport:
-    completed: bool
-    active_run_id_before_stop: str | None
-    analysis_queue_depth: int
-    analysis_active_run_id: str | None
-    analysis_queue_oldest_age_s: float | None
-    analysis_in_progress: bool
-    write_error: str | None
-    final_status: dict[str, object]
-
-
-def _build_run_metadata_record(
-    recorder: RunRecorder,
-    run_id: str,
-    start_time_utc: str,
-) -> JsonObject:
-    return build_run_metadata(
-        run_id=run_id,
-        start_time_utc=start_time_utc,
-        analysis_settings_snapshot=recorder._analysis_settings_snapshot(),
-        sensor_model=recorder.sensor_model,
-        firmware_version=firmware_version_for_run(recorder.registry),
-        default_sample_rate_hz=recorder.default_sample_rate_hz,
-        metrics_log_hz=recorder.metrics_log_hz,
-        fft_window_size_samples=recorder.fft_window_size_samples,
-        accel_scale_g_per_lsb=recorder.accel_scale_g_per_lsb,
-        active_car_snapshot=(
-            recorder._settings_store.active_car_snapshot()
-            if recorder._settings_store is not None
-            else None
-        ),
-        language_provider=recorder._language_provider,
-    )
-
-
-def _shutdown_report(recorder: RunRecorder, timeout_s: float) -> RecorderShutdownReport:
-    with recorder._lock:
-        active_run_id_before_stop = recorder._run_id
-    recorder._lifecycle.shutdown_requested = True
-    try:
-        final_status = recorder.stop_recording()
-        analysis_completed = recorder.wait_for_post_analysis(timeout_s)
-        health = recorder.health_snapshot()
-        return RecorderShutdownReport(
-            completed=analysis_completed,
-            active_run_id_before_stop=active_run_id_before_stop,
-            analysis_queue_depth=health["analysis_queue_depth"],
-            analysis_active_run_id=health["analysis_active_run_id"],
-            analysis_queue_oldest_age_s=health["analysis_queue_oldest_age_s"],
-            analysis_in_progress=health["analysis_in_progress"],
-            write_error=health["write_error"],
-            final_status=final_status,
-        )
-    finally:
-        recorder._lifecycle.shutdown_requested = False
-
-
-async def _run_loop(recorder: RunRecorder) -> None:
-    interval = 1.0 / recorder.metrics_log_hz
-    while True:
-        try:
-            timestamp_utc = utc_now_iso()
-            with recorder._lock:
-                live_start = recorder._live_start_mono_s
-                run_id_for_live = recorder._run_id or "live"
-            live_t_s = max(0.0, time.monotonic() - live_start)
-            live_rows = await asyncio.wait_for(
-                asyncio.to_thread(
-                    recorder._sample_flush.build_sample_records,
-                    run_id=run_id_for_live,
-                    t_s=live_t_s,
-                    timestamp_utc=timestamp_utc,
-                ),
-                timeout=_DB_THREAD_TIMEOUT_S,
-            )
-            if not isinstance(live_rows, list):
-                LOGGER.warning(
-                    "Metrics logger sample builder returned %s instead of list; dropping tick.",
-                    type(live_rows).__name__,
-                )
-                live_rows = []
-            snapshot = recorder._session_snapshot()
-            if snapshot is not None:
-                no_data_timeout = await asyncio.wait_for(
-                    asyncio.to_thread(
-                        recorder._sample_flush.append_records,
-                        snapshot.run_id,
-                        snapshot.start_time_utc,
-                        snapshot.start_mono_s,
-                        prebuilt_rows=live_rows,
-                    ),
-                    timeout=_DB_THREAD_TIMEOUT_S,
-                )
-                if no_data_timeout:
-                    LOGGER.info(
-                        "Auto-stopping run %s after %.1fs without new data",
-                        snapshot.run_id,
-                        recorder._lifecycle.no_data_timeout_s,
-                    )
-                    recorder.stop_recording(_only_if_run_id=snapshot.run_id)
-        except TimeoutError:
-            recorder._persistence.set_last_write_error("metrics logger DB call timed out")
-            LOGGER.warning(
-                "Metrics logger DB call exceeded %.1fs timeout; skipping tick.",
-                _DB_THREAD_TIMEOUT_S,
-            )
-        except Exception as exc:
-            recorder._persistence.set_last_write_error(f"metrics logger tick failed: {exc}")
-            LOGGER.warning(
-                "Metrics logger tick failed; will retry next interval.",
-                exc_info=True,
-            )
-        await asyncio.sleep(interval)
 
 
 class RunRecorder:
@@ -211,7 +55,7 @@ class RunRecorder:
 
     def __init__(
         self,
-        config: RunRecorderConfig,
+        config: _recorder_types.RunRecorderConfig,
         registry: ClientTracker,
         gps_monitor: SpeedProvider,
         processor: SignalSource,
@@ -227,32 +71,29 @@ class RunRecorder:
         self.sensor_model = config.sensor_model.strip() or "unknown"
         self.default_sample_rate_hz = int(config.default_sample_rate_hz)
         self.fft_window_size_samples = int(config.fft_window_size_samples)
-        self.accel_scale_g_per_lsb = (
-            float(config.accel_scale_g_per_lsb)
-            if isinstance(config.accel_scale_g_per_lsb, NUMERIC_TYPES)
-            and config.accel_scale_g_per_lsb > 0
-            else None
+        self.accel_scale_g_per_lsb = _recorder_runtime.normalize_accel_scale_g_per_lsb(
+            config.accel_scale_g_per_lsb,
         )
         self._lock = RLock()
         self._history_db = history_db
         self._language_provider = language_provider
         self._live_start_mono_s = time.monotonic()
 
-        # --- Session lifecycle ---
         self._lifecycle = RunLifecycleState(
             no_data_timeout_s=max(1.0, float(config.no_data_timeout_s)),
         )
 
-        # --- Persistence coordination ---
         self._persistence = RunPersistenceWriter(
             lock=self._lock,
             history_db=history_db,
             persist_history_db_enabled=config.persist_history_db,
             run_id_matches=self._run_id_matches,
-            metadata_builder=lambda run_id, start_time_utc: _build_run_metadata_record(
-                self,
-                run_id,
-                start_time_utc,
+            metadata_builder=lambda run_id, start_time_utc: (
+                _recorder_types._build_run_metadata_record(
+                    self,
+                    run_id,
+                    start_time_utc,
+                )
             ),
             monotonic=lambda: time.monotonic(),
             sleep=lambda seconds: time.sleep(seconds),
@@ -274,7 +115,7 @@ class RunRecorder:
             default_sample_rate_hz=self.default_sample_rate_hz,
             lifecycle=self._lifecycle,
             persistence=self._persistence,
-            active_frames_total=self._active_frames_total,
+            active_frames_total=lambda: _recorder_runtime.active_frames_total(self.registry),
             current_run_id=lambda: self._run_id,
             monotonic=lambda: time.monotonic(),
         )
@@ -285,14 +126,10 @@ class RunRecorder:
                 analysis_settings_snapshot=self._analysis_settings_snapshot(),
                 start_time_utc=utc_now_iso(),
                 start_mono_s=time.monotonic(),
-                current_total=self._active_frames_total(),
+                current_total=_recorder_runtime.active_frames_total(self.registry),
             )
             self._persistence.reset()
             self._live_start_mono_s = snapshot.start_mono_s
-
-    # -----------------------------------------------------------------------
-    # Properties
-    # -----------------------------------------------------------------------
 
     @property
     def enabled(self) -> bool:
@@ -315,24 +152,10 @@ class RunRecorder:
         return current is not None and current.run_id == run_id
 
     def _analysis_settings_snapshot(self) -> AnalysisSettingsSnapshot:
-        if self._settings_store is not None:
-            return self._settings_store.analysis_settings_snapshot()
-        return AnalysisSettingsSnapshot.from_dict(AnalysisSettingsSnapshot.DEFAULTS)
-
-    def _active_frames_total(self) -> int:
-        _get = self.registry.get
-        return sum(
-            int(rec.frames_total)
-            for cid in self.registry.active_client_ids()
-            if (rec := _get(cid)) is not None
-        )
+        return _recorder_runtime.analysis_settings_snapshot(self._settings_store)
 
     def _session_snapshot(self) -> ActiveRunSnapshot | None:
         return self._lifecycle.snapshot()
-
-    # -----------------------------------------------------------------------
-    # Public API
-    # -----------------------------------------------------------------------
 
     def status(self) -> dict[str, object]:
         return build_run_recorder_status(
@@ -389,7 +212,7 @@ class RunRecorder:
                 analysis_settings_snapshot=self._analysis_settings_snapshot(),
                 start_time_utc=utc_now_iso(),
                 start_mono_s=time.monotonic(),
-                current_total=self._active_frames_total(),
+                current_total=_recorder_runtime.active_frames_total(self.registry),
             )
             self._persistence.reset()
             self._live_start_mono_s = snapshot.start_mono_s
@@ -445,11 +268,11 @@ class RunRecorder:
     def wait_for_post_analysis(self, timeout_s: float = 30.0) -> bool:
         return self._post_analysis.wait(timeout_s)
 
-    def shutdown_report(self, timeout_s: float = 30.0) -> RecorderShutdownReport:
-        return _shutdown_report(self, timeout_s)
+    def shutdown_report(self, timeout_s: float = 30.0) -> _recorder_types.RecorderShutdownReport:
+        return _recorder_types._shutdown_report(self, timeout_s)
 
     def shutdown(self, timeout_s: float = 30.0) -> bool:
         return self.shutdown_report(timeout_s).completed
 
     async def run(self) -> None:
-        await _run_loop(self)
+        await _recorder_runtime.run_loop(self, logger=LOGGER)
